+++ server/src/bruFiles/completions/mapEnvVariablesToCompletions.ts
@@ -1,14 +1,29 @@
 import {
-    groupReferencesByName,
     VariableReferenceType,
     Range,
     BrunoVariableReference,
+    Position,
 } from "@global_shared";
 import {
     EnvVariableCommonRequestData,
+    groupReferencesByName,
     mapStaticEnvVariablesToCompletions,
+    ReferenceFromOwnFileDetails,
 } from "../../shared";
-import { CompletionItem, CompletionItemKind } from "vscode-languageserver";
+import {
+    CompletionItem,
+    CompletionItemKind,
+    TextEdit,
+} from "vscode-languageserver";
+import { EquivalentDynamicReferencesFromOtherFiles } from "../shared/interfaces";
+
+interface MatchingDynamicEnvVariables {
+    fromSameFile: {
+        blockName: string;
+        variableReference: BrunoVariableReference;
+    }[];
+    fromOtherFiles: EquivalentDynamicReferencesFromOtherFiles[];
+}
 
 export function mapEnvVariablesToCompletions(
     matchingStaticEnvVariables: {
@@ -16,82 +31,240 @@ export function mapEnvVariablesToCompletions(
         matchingVariableKeys: string[];
         isConfiguredEnv: boolean;
     }[],
-    matchingDynamicEnvVariables: {
-        blockName: string;
-        variableReference: BrunoVariableReference;
-    }[],
+    matchingDynamicEnvVariables: MatchingDynamicEnvVariables,
     requestData: EnvVariableCommonRequestData,
     appendOnInsertion?: string,
 ) {
+    const resultsForDynamicVariables = mapDynamicEnvVariables(
+        requestData,
+        matchingDynamicEnvVariables,
+        {
+            prefixForSortText: "a",
+            appendOnInsertion,
+        },
+    );
+
     const resultsForStaticVariables = mapStaticEnvVariablesToCompletions(
         requestData,
-        matchingStaticEnvVariables,
+        filterOutStaticVariablesWithDynamicReferences(
+            matchingStaticEnvVariables,
+            matchingDynamicEnvVariables,
+        ),
         // Display static environment variables below dynamic ones.
         { prefixForSortText: "b", appendOnInsertion },
     );
 
-    return resultsForStaticVariables.concat(
-        mapDynamicEnvVariables(requestData, matchingDynamicEnvVariables, {
-            prefixForSortText: "a",
-            appendOnInsertion,
-        }).filter(
-            ({ label }) =>
-                !matchingStaticEnvVariables
-                    .flatMap(({ matchingVariableKeys }) => matchingVariableKeys)
-                    .some((key) => key == label),
-        ),
-    );
+    return resultsForDynamicVariables.concat(resultsForStaticVariables);
 }
 
 function mapDynamicEnvVariables(
     requestData: EnvVariableCommonRequestData,
-    matchingDynamicEnvVariables: {
-        blockName: string;
-        variableReference: BrunoVariableReference;
-    }[],
+    { fromSameFile, fromOtherFiles }: MatchingDynamicEnvVariables,
     modifications: {
         prefixForSortText: string;
         appendOnInsertion?: string;
     },
 ) {
+    const groupedRefs = groupReferencesByName(fromSameFile, fromOtherFiles);
+
+    return groupedRefs
+        .map((ref) => {
+            if (
+                ref.hasReferenceInOwnFile &&
+                ref.detailsForOwnFileRefs != undefined
+            ) {
+                return getCompletionForRefsWithinOwnFile(
+                    requestData,
+                    {
+                        ...ref,
+                        detailsForOwnFileRefs: ref.detailsForOwnFileRefs,
+                    },
+                    modifications,
+                );
+            }
+
+            if (ref.referencesFromOtherFiles != undefined) {
+                return getCompletionForRefsFromOnlyOtherFiles(
+                    requestData,
+                    {
+                        ...ref,
+                        referencesFromOtherFiles: ref.referencesFromOtherFiles,
+                    },
+                    modifications,
+                );
+            }
+
+            return undefined;
+        })
+        .filter((v) => v != undefined);
+}
+
+function getCompletionForRefsWithinOwnFile(
+    { variable: { start, end } }: EnvVariableCommonRequestData,
+    groupedReferences: {
+        variableName: string;
+        referenceType: VariableReferenceType;
+        detailsForOwnFileRefs: ReferenceFromOwnFileDetails;
+        referencesFromOtherFiles?: EquivalentDynamicReferencesFromOtherFiles;
+    },
+    modifications: {
+        prefixForSortText: string;
+        appendOnInsertion?: string;
+    },
+): CompletionItem {
     const {
-        variable: { start, end },
-    } = requestData;
-
-    return groupReferencesByName(matchingDynamicEnvVariables).map(
-        ({
+        detailsForOwnFileRefs: {
             blockName,
-            variableName,
-            referenceType,
-            references: {
-                distinctBlocks,
-                hasDuplicateReferences,
-                totalNumberOfReferences,
-            },
-        }) => {
-            const completionItem: CompletionItem = {
-                label: variableName,
-                labelDetails: {
-                    description:
-                        hasDuplicateReferences && distinctBlocks.length > 1
-                            ? `  Blocks '${distinctBlocks.join("','")}'`
-                            : `  Block '${blockName}'`,
-                },
-                kind:
-                    referenceType == VariableReferenceType.Read
-                        ? CompletionItemKind.Field
-                        : CompletionItemKind.Function,
-                detail: hasDuplicateReferences
-                    ? `Found a total of ${totalNumberOfReferences} relevant references in ${distinctBlocks.length > 1 ? `blocks ${JSON.stringify(distinctBlocks)}` : `block '${blockName}'`}.`
-                    : undefined,
-                sortText: `${modifications?.prefixForSortText ?? ""}_${blockName}_${variableName}`,
-                textEdit: {
-                    newText: `${variableName}${modifications.appendOnInsertion ?? ""}`,
-                    range: new Range(start, end),
-                },
-            };
-
-            return completionItem;
+            allDistinctBlocks,
+            hasDuplicateReferences,
+            totalNumberOfReferences: numberOfRefsWithinFile,
         },
+        referenceType,
+        variableName,
+        referencesFromOtherFiles,
+    } = groupedReferences;
+
+    const totalNumberOfReferences =
+        numberOfRefsWithinFile +
+        (referencesFromOtherFiles?.otherMatchingReferences.length ?? -1) +
+        1;
+
+    return {
+        label: variableName,
+        labelDetails: {
+            description:
+                hasDuplicateReferences && allDistinctBlocks.length > 1
+                    ? `Blocks '${allDistinctBlocks.join("','")}'`
+                    : `Block '${blockName}'`,
+        },
+        kind: getKind(referenceType),
+        detail:
+            !hasDuplicateReferences && referencesFromOtherFiles == undefined
+                ? undefined
+                : `${totalNumberOfReferences} relevant references in ${allDistinctBlocks.length > 1 ? `blocks ${JSON.stringify(allDistinctBlocks)}` : `block '${blockName}'`}.`.concat(
+                      referencesFromOtherFiles == undefined
+                          ? ""
+                          : ` and in ${referencesFromOtherFiles.otherMatchingReferences.length + 1} other file(s)`,
+                  ),
+        sortText: getSortText(
+            modifications.prefixForSortText,
+            variableName,
+            0,
+            blockName,
+        ),
+        textEdit: getTextEdit(
+            variableName,
+            start,
+            end,
+            modifications.appendOnInsertion,
+        ),
+    };
+}
+
+function getCompletionForRefsFromOnlyOtherFiles(
+    { variable: { start, end } }: EnvVariableCommonRequestData,
+    groupedReferences: {
+        variableName: string;
+        referenceType: VariableReferenceType;
+        referencesFromOtherFiles: EquivalentDynamicReferencesFromOtherFiles;
+    },
+    modifications: {
+        prefixForSortText: string;
+        appendOnInsertion?: string;
+    },
+): CompletionItem {
+    const {
+        referenceType,
+        variableName,
+        referencesFromOtherFiles: {
+            mostRelevantReference,
+            otherMatchingReferences,
+        },
+    } = groupedReferences;
+
+    return {
+        label: variableName,
+        labelDetails: {
+            description: `${mostRelevantReference.relativePathToSourceFile}`,
+        },
+        kind: getKind(referenceType),
+        detail:
+            otherMatchingReferences.length == 0
+                ? undefined
+                : `Relevant reference(s) in ${otherMatchingReferences.length} other file(s).`,
+        sortText: getSortText(
+            modifications.prefixForSortText,
+            variableName,
+            mostRelevantReference.indirectionLevel,
+        ),
+        textEdit: getTextEdit(
+            variableName,
+            start,
+            end,
+            modifications.appendOnInsertion,
+        ),
+    };
+}
+
+function filterOutStaticVariablesWithDynamicReferences(
+    matchingStaticEnvVariables: {
+        environmentFile: string;
+        matchingVariableKeys: string[];
+        isConfiguredEnv: boolean;
+    }[],
+    matchingDynamicEnvVariables: MatchingDynamicEnvVariables,
+) {
+    const allVariableNamesFromDynamicReferences =
+        matchingDynamicEnvVariables.fromSameFile
+            .map(({ variableReference: { variableName } }) => variableName)
+            .concat(
+                matchingDynamicEnvVariables.fromOtherFiles.map(
+                    ({
+                        mostRelevantReference: {
+                            reference: { variableName },
+                        },
+                    }) => variableName,
+                ),
+            );
+
+    return matchingStaticEnvVariables.map(
+        ({
+            environmentFile,
+            isConfiguredEnv,
+            matchingVariableKeys: allMatchingKeys,
+        }) => ({
+            environmentFile,
+            isConfiguredEnv,
+            matchingVariableKeys: allMatchingKeys.filter(
+                (key) => !allVariableNamesFromDynamicReferences.includes(key),
+            ),
+        }),
     );
+}
+
+function getKind(referenceType: VariableReferenceType) {
+    return referenceType == VariableReferenceType.Read
+        ? CompletionItemKind.Field
+        : CompletionItemKind.Function;
+}
+
+function getTextEdit(
+    variableName: string,
+    start: Position,
+    end: Position,
+    appendOnInsertion?: string,
+): TextEdit {
+    return {
+        newText: `${variableName}${appendOnInsertion ?? ""}`,
+        range: new Range(start, end),
+    };
+}
+
+function getSortText(
+    prefixForSortText: string,
+    variableName: string,
+    indirectionLevel: number,
+    blockName?: string,
+) {
+    return `${prefixForSortText ?? ""}_${indirectionLevel}_${variableName}${blockName ?? ""}`;
 }