+++ shared/baseModel/typeguards/isCollectionDirectory.ts
@@ -1,0 +1,11 @@
+import {
+    CollectionDirectory,
+    CollectionItem,
+    isCollectionItemWithSequence,
+} from "../../";
+
+export function isCollectionDirectory(
+    item: CollectionItem,
+): item is CollectionDirectory {
+    return !item.isFile() && isCollectionItemWithSequence(item);
+}